rearrange code to be more clear
